+++ R/table_as_data_frame.R
@@ -1,4 +1,6 @@
 unlist_n <- function(x, times) {
+  if (times <= 0)
+    return(x)
   x <- unlist(x, recursive = FALSE)
   if (times == 1)
     return(x)
@@ -28,7 +30,7 @@ sc_model_matrix <- function(dims) {
       rep(times = times, each = each)
     times <- times * dims[i]
   }
-  as.data.frame(out)
+  vctrs::new_data_frame(out)
 }
 
 sc_table_create_data <- function(content) {